Update README.md
I was testing this process on a Windows dev machine, and without the cp command available, I needed to use copy with \ instead of /. Also, unless node.exe is the default processor for .js files, which it wasn't on my machine, it's necessary to run r.js.cmd instead of just r.js. With my default node installation, r.js is found in that folder before r.js.cmd, so adding .cmd is necessary to disambiguate.